preparing mgmesytec for unit testing
