+++ prototype2/multigrid/MGMesytecBase.cpp
@@ -7,31 +7,18 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include <common/Detector.h>
-#include <common/EFUArgs.h>
+#include <multigrid/MGMesytecBase.h>
+
 #include <common/EV42Serializer.h>
 #include <common/Producer.h>
-#include <common/RingBuffer.h>
 #include <common/TimeString.h>
-#include <cstring>
 #include <efu/Parser.h>
 #include <efu/Server.h>
-#include <common/HistSerializer.h>
-#include <common/ReadoutSerializer.h>
-#include <iostream>
-#include <libs/include/SPSCFifo.h>
+
 #include <libs/include/Socket.h>
 #include <libs/include/TSCTimer.h>
 #include <libs/include/Timer.h>
-#include <memory>
 #include <multigrid/mgmesytec/Sis3153Parser.h>
-#include <multigrid/mgmesytec/Vmmr16Parser.h>
-#include <multigrid/MgConfig.h>
-#include <multigrid/mgmesytec/HitFile.h>
-#include <queue>
-#include <stdio.h>
-#include <unistd.h>
-
 #include <multigrid/mgmesytec/SequoiaGeometry.h>
 //#include <multigrid/mgmesytec/MG24Geometry.h>
 
@@ -39,138 +26,11 @@
 #include <multigrid/mgmesytec/EfuCenterMass.h>
 #include <multigrid/mgmesytec/EfuPrioritized.h>
 
-#include <common/Log.h>
-
 const int TSC_MHZ = 2900; // Not accurate, do not rely solely on this
 
-/** ----------------------------------------------------- */
-struct DetectorSettingsStruct {
-  std::string ConfigFile;
-  bool monitor{false};
-  std::string fileprefix{""};
-} DetectorSettings;
 
-void SetCLIArguments(CLI::App & parser) {
-  parser.add_option("-f,--file", DetectorSettings.ConfigFile,
-                    "NMX (gdgem) specific config file")->group("MGMesytec")->
-                    required()->configurable(true);
-  parser.add_flag("--monitor", DetectorSettings.monitor,
-                  "stream monitor data")->group("MGMesytec")->configurable(true)->default_val("true");
-  parser.add_option("--dumptofile", DetectorSettings.fileprefix,
-                    "dump to specified file")->group("MGMesytec")->configurable(true);
-}
-
-PopulateCLIParser PopulateParser{SetCLIArguments};
-
-struct Monitor
-{
-  std::shared_ptr<Hists> hists;
-  std::shared_ptr<ReadoutSerializer> readouts;
-
-  void init(std::string broker, size_t max_readouts) {
-    readouts = std::make_shared<ReadoutSerializer>(max_readouts);
-    hists = std::make_shared<Hists>(std::numeric_limits<uint16_t>::max(),
-                                    std::numeric_limits<uint16_t>::max());
-    histfb = std::make_shared<HistSerializer>(hists->needed_buffer_size());
-
-    producer = std::make_shared<Producer>(broker, "C-SPEC_monitor");
-    readouts->set_callback(std::bind(&Producer::produce2<uint8_t>, producer.get(), std::placeholders::_1));
-    histfb->set_callback(std::bind(&Producer::produce2<uint8_t>, producer.get(), std::placeholders::_1));
-    enabled_ = true;
-  }
-
-  void close() {
-    enabled_ = false;
-    hists.reset();
-    readouts.reset();
-    histfb.reset();
-    producer.reset();
-  }
-
-  void produce() {
-    if (!enabled_)
-      return;
-
-    if (!hists->isEmpty()) {
-      LOG(PROCESS, Sev::Debug, "Flushing histograms for {} readouts", hists->hit_count());
-      histfb->produce(*hists);
-      hists->clear();
-    }
-
-    if (readouts->getNumEntries()) {
-      LOG(PROCESS, Sev::Debug, "Flushing readout data for {} readouts", readouts->getNumEntries());
-      readouts->produce();
-    }
-  }
-
-  bool enabled() const {
-    return enabled_;
-  }
-
-  Monitor() = default;
-  ~Monitor() { close(); }
-
-private:
-  bool enabled_ {false};
-
-  std::shared_ptr<Producer> producer;
-  std::shared_ptr<HistSerializer> histfb;
-};
-
-/** ----------------------------------------------------- */
-const char *classname = "CSPEC Detector Mesytec readout";
-
-///
-class CSPEC : public Detector {
-public:
-  explicit CSPEC(BaseSettings settings);
-  void mainThread();
-
-  const char *detectorname();
-
-  /// Some hardcoded constants
-  static constexpr int eth_buffer_size = 9000;          /// used for experimentation
-  const int kafka_buffer_size = 1000000;     /// -||-
-  const int readout_entries = 100000;        /// number of raw readout entries
-  const int one_tenth_second_usecs = 100000; ///
-
-private:
-
-  struct {
-    // Input Counters
-    int64_t rx_packets {0};
-    int64_t rx_bytes {0};
-    int64_t sis_discarded_bytes {0};
-    int64_t vmmr_discarded_bytes {0};
-    int64_t triggers {0};
-    int64_t bus_glitches {0};
-    int64_t bad_triggers {0};
-    int64_t readouts {0};
-    int64_t readouts_discarded {0};
-    int64_t readouts_culled {0};
-    int64_t geometry_errors {0};
-    int64_t timing_errors {0};
-    int64_t events {0};
-    int64_t tx_bytes {0};
-  } ALIGN(64) mystats;
-
-  void init_config();
-
-  Multigrid::Config mg_config;
-  Monitor monitor;
-
-  bool HavePulseTime{false};
-  uint64_t RecentPulseTime{0};
-
-  uint64_t ShortestPulsePeriod{std::numeric_limits<uint64_t>::max()};
-
-  Multigrid::VMMR16Parser vmmr16Parser;
-
-  std::shared_ptr<Multigrid::Efu> mgEfu;
-  std::shared_ptr<Multigrid::HitFile> dumpfile;
-};
-
-CSPEC::CSPEC(BaseSettings settings) : Detector("CSPEC", settings) {
+MGMesytecBase::MGMesytecBase(BaseSettings const &settings, struct MGMesytecSettings & LocalSettings)
+       : Detector("CSPEC", settings), MGMesytecSettings(LocalSettings) {
   Stats.setPrefix("efu.mgmesytec");
 
   LOG(INIT, Sev::Info, "Adding stats");
@@ -191,27 +51,24 @@ CSPEC::CSPEC(BaseSettings settings) : Detector("CSPEC", settings) {
   Stats.create("tx_bytes",             mystats.tx_bytes);
   // clang-format on
 
-  std::function<void()> inputFunc = [this]() { CSPEC::mainThread(); };
+  std::function<void()> inputFunc = [this]() { MGMesytecBase::mainThread(); };
   Detector::AddThreadFunction(inputFunc, "main");
 
   LOG(INIT, Sev::Info, "Stream monitor data = {}",
-      (DetectorSettings.monitor ? "YES" : "no"));
-  if (!DetectorSettings.fileprefix.empty())
+      (MGMesytecSettings.monitor ? "YES" : "no"));
+  if (!MGMesytecSettings.fileprefix.empty())
     LOG(INIT, Sev::Info, "Dump h5 data in path: {}",
-           DetectorSettings.fileprefix);
+           MGMesytecSettings.fileprefix);
 }
 
-const char *CSPEC::detectorname() { return classname; }
-
-
-
-void CSPEC::init_config()
+///
+void MGMesytecBase::init_config()
 {
-  LOG(INIT, Sev::Info, "MG Config file: {}", DetectorSettings.ConfigFile);
-  mg_config = Multigrid::Config(DetectorSettings.ConfigFile);
+  LOG(INIT, Sev::Info, "MG Config file: {}", MGMesytecSettings.ConfigFile);
+  mg_config = Multigrid::Config(MGMesytecSettings.ConfigFile);
   LOG(INIT, Sev::Info, "Multigrid Config\n{}", mg_config.debug());
 
-  if (DetectorSettings.monitor)
+  if (MGMesytecSettings.monitor)
     monitor.init(EFUSettings.KafkaBroker, readout_entries);
 
   if (mg_config.reduction_strategy == "center-mass") {
@@ -226,15 +83,15 @@ void CSPEC::init_config()
   mgEfu->hists = monitor.hists;
   mgEfu->raw1 = monitor.readouts;
 
-  if (!DetectorSettings.fileprefix.empty())
+  if (!MGMesytecSettings.fileprefix.empty())
   {
     dumpfile = Multigrid::HitFile::create(
-        DetectorSettings.fileprefix + "mgmesytec_" + timeString(), 100);
+        MGMesytecSettings.fileprefix + "mgmesytec_" + timeString(), 100);
   }
   vmmr16Parser.spoof_high_time(mg_config.spoof_high_time);
 }
 
-void CSPEC::mainThread() {
+void MGMesytecBase::mainThread() {
   init_config();
 
   /** Connection setup */
@@ -342,7 +199,3 @@ void CSPEC::mainThread() {
     }
   }
 }
-
-/** ----------------------------------------------------- */
-
-DetectorFactory<CSPEC> Factory;
